Create 2021-6-08-Session-2.md
---
layout: post
title: "Session #2 - Crouching Owlbear, Hidden Dragon"
categories: session-recaps
---
With a leather satchel full of medical supplies in tow, the party sets off for the far town of Woodhaerst during their journey to the Woodland Corps home base. Their passage through the Kinderkine Woods is, for the most part, an uneventful trek as the geography begins to wrinkle into large, stony hills dressed in thick vegetation.

Our small, mink friend Sylphrena sniffs under a tree and finds a boot -- but not just any boot, a gross bloody one with a little bit of foot left inside! This prompts the party to investigate their surroundings and find further evidence of struggle and violence nearby. Though he does not know what they are, Sulu notices a very large number of horse tracks on the path ahead of them. 

More evidence reveals itself -- a mess of knocked over trees and the words ‘GO AWAY’ dug into the dirt reveal to the party that something, like before, is ‘a foot’. 

Outside the village gates the party is approached by a large Owlbear (not a cyclops) named Po who claims to be from the Woodland Corps home base and is here to assist in their arrival. At the time of meeting them, Po was eating a mouse. 

The party progresses inside much to the fear and terror of those who see them, including one Aywin Bradley who had stolen her father’s sword and helmet to fend off the invading Team 287 herself.

As war horns ring and a stampede of hooves rushes down the hillside, the group sprints, dives, and climbs their way into a veritable medley of hidey holes -- Including trees, hay piles, wells, and even homes made for tiny people. 

From the shadows, the party watches as a group of bandits lead by a large centaur stumble into Woodhaerst and begin shaking down its residents. This sight picks a particular chord with Veera, as she jumps out of the pig pen and challenges the cowardly bandits to throw down. 

The ensuing battle sees the full breadth of the party’s many years of training at the academy, as firebolts, devastating slashes, and drop bear spec op tactics tear through the bandit attackers (literally, in Kaladin’s case.)

Kaladin and Veera each manage to capture a bandit, though Kaladin is too much of a bear to get anything more than a little bit of pee out of his captive. Veera, Sulu, and Maria stand over the terrified bandit (named Ultimus) and from him are able to gleen the rough number of remaining bandits as well as the location of their hideout. Veera then insists that this man changes his ways and venture ahead on a new path -- one that leads to the Academy gates so that he may himself enroll.

And much to the chagrin of the new bald man in the sky, that tall, antlered buttmuncher took my goddamn plothook away from me, you big, teleporting fox having motherfucker. 

Yuka and Podrick begin walking back as Riona makes threats of violent retaliation.